requirements.txt
# =========================
# CORE WEB FRAMEWORK
# =========================
streamlit==1.41.1
fastapi==0.115.5
uvicorn==0.32.1

# =========================
# ENV MANAGEMENT
# =========================
python-dotenv==1.0.1

# =========================
# LLM PROVIDERS
# =========================
mistralai==0.4.3
google-generativeai==0.8.6
openai==1.56.0

# =========================
# LANGCHAIN & CORE
# =========================
langchain==0.2.16
langchain-core==0.2.47
langchain-community==0.2.16
langchain-text-splitters==0.2.2

# LLM INTEGRATIONS
langchain-mistralai==0.1.19
langchain-google-genai==0.0.14
langchain-openai==0.1.35

# =========================
# VECTOR DATABASES (RAG MEMORY)
# =========================
chromadb==1.5.7
faiss-cpu==1.10.0

# =========================
# EMBEDDINGS & DOCUMENT PROCESSING
# =========================
sentence-transformers==3.4.1
pypdf==4.3.1
beautifulsoup4==4.14.3
python-docx==1.0.1
unstructured==0.15.33

# =========================
# UTILITIES
# =========================
requests==2.33.0
aiohttp==3.13.5
typing-extensions==4.12.2

# =========================
# DATA HANDLING
# =========================
numpy
pandas

# =========================
# HTTP / API HANDLING
# =========================
requests
httpx

# =========================
# UTILITIES
# =========================
tqdm
tenacity

# =========================
# TOKENIZATION / TEXT SPLITTING SUPPORT
# =========================
tiktoken

# =========================
# OPTIONAL (ADVANCED RAG FEATURES)
# =========================
sqlalchemy